Move tiles hooks to their own dir
