+++ src/hooks/tile-game/useTileAsset.ts
@@ -1,5 +1,5 @@
 import { v4 } from 'uuid';
-import { TileAsset as TileAssetT } from '../types';
+import { TileAsset as TileAssetT } from '../../types';
 
 export interface TileAssetI {
   create: (image: string) => TileAssetT;
